requirements.txt
Django==5.1
pillow==10.4.0
crispy-bootstrap5==2024.2
djangorestframework==3.15.2
django-environ==0.11.2